+++ ecommerceComponent/index.html
@@ -103,7 +103,7 @@
       </div>
     </div>
   </div>
-  <script src="app.js"></script>
+  <script src="app.js" type ="module"></script>
 </body>
 
 </html>
